refactor(Piece): 중복 로직 상위로 이동
