+++ src/main/java/chess/domain/piece/Bishop.java
@@ -1,11 +1,8 @@
 package chess.domain.piece;
 
-import chess.domain.Position;
 import chess.domain.piece.character.Character;
 import chess.domain.piece.character.Kind;
 import chess.domain.piece.character.Team;
-import java.util.ArrayList;
-import java.util.List;
 
 public class Bishop extends Piece {
     public Bishop(Team team) {
@@ -25,25 +22,12 @@ public class Bishop extends Piece {
     }
 
     @Override
-    public Character findCharacter() {
-        return Character.findCharacter(team, Kind.BISHOP);
-    }
-
-    @Override
-    protected List<Position> findBetweenPositions(Position position, int rowDifference, int columnDifference) {
-        int absoluteDifference = Math.abs(rowDifference);
-        int rowSign = rowDifference / absoluteDifference;
-        int columnSign = columnDifference / absoluteDifference;
-
-        List<Position> positions = new ArrayList<>();
-        for (int movement = MIN_MOVEMENT; movement < absoluteDifference; movement++) {
-            positions.add(position.move(rowSign * movement, columnSign * movement));
-        }
-        return positions;
-    }
-
-    @Override
     protected boolean isMovable(int rowDifference, int columnDifference) {
         return Math.abs(rowDifference) == Math.abs(columnDifference);
+    }
+
+    @Override
+    public Character findCharacter() {
+        return Character.findCharacter(team, Kind.BISHOP);
     }
 }
