+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,12 +1,8 @@
 package chess.domain.piece;
 
-import chess.Calculator;
-import chess.domain.Position;
 import chess.domain.piece.character.Character;
 import chess.domain.piece.character.Kind;
 import chess.domain.piece.character.Team;
-import java.util.ArrayList;
-import java.util.List;
 
 public class Queen extends Piece {
 
@@ -29,19 +25,6 @@ public class Queen extends Piece {
     @Override
     public Character findCharacter() {
         return Character.findCharacter(team, Kind.QUEEN);
-    }
-
-    @Override
-    protected List<Position> findBetweenPositions(Position position, int rowDifference, int columnDifference) {
-        int absoluteDifference = Math.max(Math.abs(rowDifference), Math.abs(columnDifference));
-        int rowSign = Calculator.calculateSign(rowDifference);
-        int columnSign = Calculator.calculateSign(columnDifference);
-
-        List<Position> positions = new ArrayList<>();
-        for (int movement = MIN_MOVEMENT; movement < absoluteDifference; movement++) {
-            positions.add(position.move(rowSign * movement, columnSign * movement));
-        }
-        return positions;
     }
 
     @Override
